Create regis.html
It is the registration form created with html, here API is used for getting weather for the particular location given by user. Here Visual Crossing  is used for API, from there API key ("47DK7JBNEQ26XE2C5YRJPM2L3")is taken. and with the help of this API key weather for particular location is obtained. connect.php is used to connect mySQL and html.